DBZ-8956 move in items
