+++ quarkus-debezium-parent/quarkus-debezium-engine-spi-parent/deployment/src/main/java/io/quarkus/debezium/deployment/items/DebeziumConnectorBuildItem.java
@@ -4,7 +4,7 @@
  * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
 
-package io.quarkus.debezium.deployment;
+package io.quarkus.debezium.deployment.items;
 
 import io.debezium.runtime.ConnectorProducer;
 import io.quarkus.builder.item.MultiBuildItem;
